feat: chatbot refactoring
